+++ packages/ledger-icp/src/mocks/ledger.mock.ts
@@ -1,0 +1,5 @@
+import { AccountIdentifier } from "../account_identifier";
+
+export const mockAccountIdentifier = AccountIdentifier.fromHex(
+  "3e8bbceef8b9338e56a1b561a127326e6614894ab9b0739df4cc3664d40a5958",
+);